API reference

Docstrings

GoogleHashCode2014.GoogleHashCode2014 — Module

GoogleHashCode2014

Lightweight package designed to interact with the data of the 2014 Google Hash Code.

GoogleHashCode2014.City — Type

City

Store a city made of Junctions and Streets, along with additional instance parameters.

A City object can be constructed with the function read_city.

Fields

total_duration::Int: total time allotted for the car itineraries (in seconds)

nb_cars::Int: number of cars in the fleet

starting_junction::Int: index of the junction at which all the cars are located initially (the index refers to a position in the list of junctions)

junctions::Vector{Junction}: list of junctions

streets::Vector{Street}: list of streets

Example

julia> using GoogleHashCode2014

julia> city = read_city()
City with 11348 junctions and 17958 streets.
8 cars all start from junction 4517 and travel for at most 54000s.
    
julia> length(city.junctions)
11348
    
julia> length(city.streets)
17958

julia> city.starting_junction
4517

julia> city.junctions[10]  # get the junction object at junction index 10
Junction at coordinates (48.872250900000004, 2.3124588)

julia> city.streets[10]  # get the street object at street index 10
Bidirectional street between junction indices 6814 and 2728 - duration 13s, distance 187m

GoogleHashCode2014.Junction — Type

Junction

Store a junction from a City.

Tip

Junctions are usually referred to by their integer index in a city's list of junctions. The junction object itself is mainly useful for geographic visualization.

Fields

latitude::Float64: latitude (in decimal degrees)

longitude::Float64: longitude (in decimal degrees)

Example

julia> using GoogleHashCode2014

julia> city = read_city();

julia> junction = city.junctions[10]  # get the junction object at junction index 10
Junction at coordinates (48.872250900000004, 2.3124588)

julia> junction.latitude
48.872250900000004

GoogleHashCode2014.Solution — Type

Solution

Store a list of itineraries through a City, one for each car in.

A naive Solution can be computed with the function random_walk.

Fields

itineraries::Vector{Vector{Int}}: each itinerary is a vector of junction indices

Example

julia> using GoogleHashCode2014, Random

julia> city = read_city();

julia> rng = Random.MersenneTwister(0);

julia> solution = random_walk(rng, city)
Solution with 8 itineraries of lengths [3810, 3277, 3779, 3278, 3451, 3697, 4366, 3707]

julia> length(solution.itineraries[2])  # number of junctions visited by car nb 2
3277

julia> solution.itineraries[2][1:10]  # first 10 junction indices visited by car nb 2
10-element Vector{Int64}:
  4517
  1033
  3656
  7681
   398
  4680
 10361
 10089
 10361
 10089

GoogleHashCode2014.Street — Type

Street

Store an edge between two junctions in a City.

Tip

A Street refers to junctions with their integer indices, not with the Junction object itself.

Fields

endpointA::Int: index of the first junction

endpointB::Int: index of the second junction

bidirectional::Bool: whether B -> A is allowed

duration::Int: time cost of traversing the street (in seconds)

distance::Int: length of the street (in meters)

Example

julia> using GoogleHashCode2014

julia> city = read_city();

julia> street = city.streets[10]
Bidirectional street between junction indices 6814 and 2728 - duration 13s, distance 187m

julia> street.bidirectional
true

julia> (street.endpointA, street.endpointB)
(6814, 2728)

julia> (street.distance / street.duration) * 3.6  # speed on that street in km/h
51.784615384615385

julia> street = city.streets[11]
Monodirectional street from junction index 3779 to index 7853 - duration 12s, distance 88m

julia> street.bidirectional
false

GoogleHashCode2014.change_duration — Method

change_duration(city::City, total_duration::Integer)

Return a new City with a different total_duration and every other field equal to the ones in city.

Example

julia> using GoogleHashCode2014

julia> city = read_city()
City with 11348 junctions and 17958 streets.
8 cars all start from junction 4517 and travel for at most 54000s.

julia> change_duration(city, 3600)
City with 11348 junctions and 17958 streets.
8 cars all start from junction 4517 and travel for at most 3600s.

GoogleHashCode2014.get_street_end — Method

get_street_end(i::Integer, street::Street)

Retrieve the junction index at the other end street when it is crossed starting from junction index i.

If i is not a valid starting point for street, an error is thrown.

Example

julia> using GoogleHashCode2014

julia> city = read_city();

julia> street = city.streets[10]
Bidirectional street between junction indices 6814 and 2728 - duration 13s, distance 187m

julia> get_street_end(6814, street)
2728

julia> get_street_end(2728, street)
6814

julia> street = city.streets[11]
Monodirectional street from junction index 3779 to index 7853 - duration 12s, distance 88m

julia> get_street_end(3779, street)
7853

GoogleHashCode2014.is_feasible — Method

is_feasible(solution::Solution, city::City; verbose::Bool=false)

Check if solution satisfies the constraints for the instance defined by city.

The following criteria are considered (taken from the problem statement):

the number of itineraries has to match the number of cars of city

the first junction of each itinerary has to be the starting junction of city

for each consecutive pair of junctions on an itinerary, a street connecting these junctions has to exist in city (if the street is one directional, it has to be traversed in the correct direction)

the duration of each itinerary has to be lower or equal to the total duration of city

Example

julia> using GoogleHashCode2014, Random

julia> city = read_city();

julia> rng = Random.MersenneTwister(0);

julia> solution = random_walk(rng, city)
Solution with 8 itineraries of lengths [3810, 3277, 3779, 3278, 3451, 3697, 4366, 3707]

julia> is_feasible(solution, city; verbose=false)
true

julia> partial_solution = Solution(solution.itineraries[1:7])
Solution with 7 itineraries of lengths [3810, 3277, 3779, 3278, 3451, 3697, 4366]

julia> is_feasible(partial_solution, city; verbose=false)
false

GoogleHashCode2014.is_street — Method

is_street(i::Integer, j::Integer, street::Street)

Check if the trip from junction index i to junction index j corresponds to a valid direction of street.

The valid directions for a Street are:

from street.endpointA to street.endpointB in every case

from street.endpointB to street.endpointA only if street.bidirectional is true

Example

julia> using GoogleHashCode2014

julia> city = read_city(); 

julia> street = city.streets[10]
Bidirectional street between junction indices 6814 and 2728 - duration 13s, distance 187m

julia> is_street(1, 2728, street)
false

julia> is_street(6814, 2728, street)
true

julia> is_street(2728, 6814, street)
true

julia> street = city.streets[11]
Monodirectional street from junction index 3779 to index 7853 - duration 12s, distance 88m

julia> is_street(3779, 7853, street)
true

julia> is_street(7853, 3779, street)  # because it is not bidirectional
false

GoogleHashCode2014.is_street_start — Method

is_street_start(i::Integer, street::Street)

Check if junction index i is a valid starting point of street.

Example

julia> using GoogleHashCode2014

julia> city = read_city();

julia> street = city.streets[10]
Bidirectional street between junction indices 6814 and 2728 - duration 13s, distance 187m

julia> is_street_start(6814, street)
true

julia> is_street_start(2728, street)
true

julia> street = city.streets[11]
Monodirectional street from junction index 3779 to index 7853 - duration 12s, distance 88m

julia> is_street_start(3779, street)
true

julia> is_street_start(7853, street)
false

GoogleHashCode2014.plot_streets — Function

plot_streets(
    city::City, solution::Union{Solution,Nothing}=nothing;
    path::Union{AbstractString,Nothing}=nothing
)

Plot a City and an optional Solution using the Python library folium, save the result as an HTML file at path

This method is defined in a package extension and requires PythonCall.jl to be loaded.

GoogleHashCode2014.random_walk — Method

random_walk(rng::Random.AbstractRNG, city::City)
random_walk(city:::City)

Compute and return a Solution from city by letting each car follow a random walk from its starting point.

Tip

You can pass a random number generator with a specific seed, like Random.MersenneTwister(0), to obtain reproducible results. Otherwise, the global random number generator will be used, and the results will be different for every run.

Example

julia> using GoogleHashCode2014, Random

julia> city = read_city();

julia> rng = Random.MersenneTwister(0);

julia> solution = random_walk(rng, city)
Solution with 8 itineraries of lengths [3810, 3277, 3779, 3278, 3451, 3697, 4366, 3707]

GoogleHashCode2014.read_city — Function

read_city()
read_city(path::AbstractString)

Parse and return a City, either from the default challenge data (https://storage.googleapis.com/coding-competitions.appspot.com/HC/2014/paris_54000.txt), or from a similarly formatted text file located at path.

GoogleHashCode2014.read_solution — Method

read_solution(path)

Read and parse a Solution from a file located at path.

GoogleHashCode2014.total_distance — Method

total_distance(solution::Solution, city::City)

Compute the total distance traveled by all itineraries in solution based on the street data from city.

Warning

Streets visited several times or in both directions are only counted once.

Example

julia> using GoogleHashCode2014, Random

julia> city = read_city();

julia> rng = Random.MersenneTwister(0);

julia> solution = random_walk(rng, city)
Solution with 8 itineraries of lengths [3810, 3277, 3779, 3278, 3451, 3697, 4366, 3707]

julia> total_distance(solution, city)
752407

GoogleHashCode2014.write_city — Method

write_city(city::City, path::AbstractString)

Write city to a file located at path.

GoogleHashCode2014.write_solution — Method

write_solution(solution, path)

Write a Solution to a file located at path.

Index

GoogleHashCode2014.GoogleHashCode2014

GoogleHashCode2014.City

GoogleHashCode2014.Junction

GoogleHashCode2014.Solution

GoogleHashCode2014.Street

GoogleHashCode2014.change_duration

GoogleHashCode2014.get_street_end

GoogleHashCode2014.is_feasible

GoogleHashCode2014.is_street

GoogleHashCode2014.is_street_start

GoogleHashCode2014.plot_streets

GoogleHashCode2014.random_walk

GoogleHashCode2014.read_city

GoogleHashCode2014.read_solution

GoogleHashCode2014.total_distance

GoogleHashCode2014.write_city

GoogleHashCode2014.write_solution