API reference

HashCode2014.HashCode2014 — Module

HashCode2014

Lightweight package designed to interact with the data of the 2014 Google Hash Code.

Challenge description: https://storage.googleapis.com/coding-competitions.appspot.com/HC/2014/hashcode2014_final_task.pdf

Challenge data: https://storage.googleapis.com/coding-competitions.appspot.com/HC/2014/paris_54000.txt

Made for MIT course C25: Julia: Solving Real-World Problems with Computation

HashCode2014.City — Type

City

Store a city made of Junctions and Streets, along with additional instance parameters.

Fields

total_duration::Int: total time allotted for the car itineraries (in seconds)

nb_cars::Int: number of cars in the fleet

starting_junction::Int: junction at which all the cars are located initially

junctions::Vector{Junction}: list of junctions

streets::Vector{Street}: list of streets

HashCode2014.Junction — Type

Junction

Store a city junction.

Fields

latitude::Float64: latitude (in decimal degrees)

longitude::Float64: longitude (in decimal degrees)

HashCode2014.Solution — Type

Solution

Store a set of itineraries, one for each car.

Fields

itineraries::Vector{Vector{Int}}: each itinerary is a vector of junction indices

HashCode2014.Street — Type

Street

Store an edge between two Junctions.

Fields

endpointA::Int: index of the first junction

endpointB::Int: index of the second junction

bidirectional::Bool: whether B -> A is allowed

duration::Int: time cost of traversing the street (in seconds)

distance::Int: length of the street (in meters)

HashCode2014.get_street_end — Method

get_street_end(i, street)

Retrieve the arrival endpoint of street when it starts at junction i.

HashCode2014.is_feasible — Method

is_feasible(solution, city[; verbose=false])

Check if solution satisfies the constraints for the instance defined by city. The following criteria are considered (taken from the problem statement):

the number of itineraries has to match the number of cars of city

the first junction of each itinerary has to be the starting junction of city

for each consecutive pair of junctions on an itinerary, a street connecting these junctions has to exist in city (if the street is one directional, it has to be traversed in the correct direction)

the duration of each itinerary has to be lower or equal to the total duration of city

HashCode2014.is_street — Method

is_street(i, j, street)

Check if the trip from junction i to junction j corresponds to a valid direction of street.

HashCode2014.is_street_start — Method

is_street_start(i, street)

Check if junction i corresponds to a valid starting point of street.

HashCode2014.plot_streets — Function

plot_streets(city, solution=nothing; path=nothing)

Plot a City and an optional Solution using the Python library folium, save the result as an HTML file at path.

HashCode2014.read_city — Function

read_city(path)

Read and parse a City from a file located at path.

The default path is an artifact containing the official challenge data.

HashCode2014.read_solution — Method

read_solution(solution, path)

Read and parse a Solution from a file located at path.

HashCode2014.total_distance — Method

total_distance(solution, city)

Compute the total distance of all itineraries in solution based on the street data from city. Streets visited several times are only counted once.

HashCode2014.write_city — Method

write_city(city, path)

Write a City to a file located at path.

HashCode2014.write_solution — Method

write_solution(solution, path)

Write a Solution to a file located at path.